use same precedence logic for modules and views; only remove view path if it exists when deleting

+++ lib/spack/spack/environment.py
@@ -665,18 +665,11 @@ class Environment(object):
 
         tty.msg("Updating view at {0}".format(self._view_path))
 
-        view_specs = {}
-        for root_hash in self.concretized_order:
-            root = self.specs_by_hash[root_hash]
-            # If a spec appears as a root and as a dependency, the root always
-            # overrides the dependency
-            view_specs[root.name] = root
-            for dep in root.traverse(root=False, deptype=('link', 'run')):
-                if dep.name not in view_specs:
-                    view_specs[dep.name] = dep
+        view_specs = self._get_environment_specs()
+
         view = self.view()
         specs_without_build_deps = list()
-        for spec in view_specs.values():
+        for spec in view_specs:
             specs_without_build_deps.append(
                 spack.spec.Spec.from_dict(spec.to_dict(all_deps=False)))
         view.add_specs(*specs_without_build_deps, with_dependencies=False)
@@ -690,7 +683,8 @@ class Environment(object):
             raise SpackEnvironmentError(
                 "The environment '{0}' does not maintain a view"
                 .format(self.name))
-        shutil.rmtree(self._view_path)
+        if os.path.exists(self._view_path):
+            shutil.rmtree(self._view_path)
         self.update_view()
 
     def _add_concrete_spec(self, spec, concrete, new=True):
